# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple bank API",
        "title": "Go Bank",
        "termsOfService": "https://google.com",
        "contact": {
            "name": "David Oheji",
            "url": "https://twitter.com/ejedavy",
            "email": "ejeohejidavid@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/transfer/getTransfer/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfers"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/db.Transfer"
                        }
                    },
                    "400": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/api.ServerError"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/api.ServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/transfer/getincomingtransfers": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfers"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "to_account_id",
                        "name": "to_account_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Transfer"
                            }
                        }
                    },
                    "400": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/api.ServerError"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/api.ServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/transfer/getoutgoingtransfers": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfers"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "from_account_id",
                        "name": "from_account_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Transfer"
                            }
                        }
                    },
                    "400": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/api.ServerError"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/api.ServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/transfer/initiatetransfer": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transfers"
                ],
                "parameters": [
                    {
                        "description": "requestBody",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.InitiateTransferHandlerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/db.TransferTXResult"
                        }
                    },
                    "400": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/api.ServerError"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/api.ServerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.InitiateTransferHandlerRequest": {
            "type": "object",
            "required": [
                "amount",
                "receiver_id",
                "sender_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 1
                },
                "receiver_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "sender_id": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "api.ServerError": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "db.Account": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "owner": {
                    "type": "string"
                }
            }
        },
        "db.Entry": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "description": "Can be positive or negative",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "db.Transfer": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "This can only be positive",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "from_account_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "to_account_id": {
                    "type": "integer"
                }
            }
        },
        "db.TransferTXResult": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "receiver": {
                    "$ref": "#/definitions/db.Account"
                },
                "receiverEntry": {
                    "$ref": "#/definitions/db.Entry"
                },
                "sender": {
                    "$ref": "#/definitions/db.Account"
                },
                "senderEntry": {
                    "$ref": "#/definitions/db.Entry"
                },
                "transfer": {
                    "$ref": "#/definitions/db.Transfer"
                }
            }
        }
    }
}